chore: Improve lazy loading for images in ProductsSection
Update the ProductsSection component to improve the lazy loading of images. This change includes adding the `isFetching` state to the `useQuery` hook and displaying a loading skeleton when either `isLoading` or `isFetching` is true. This enhancement aims to provide a better user experience by optimizing the loading of product images.